Move PlaybackControl into models/
