+++ include/fullscore/widgets/gui_score_editor.h
@@ -7,8 +7,8 @@
 #include <allegro_flare/gui/widget.h>
 #include <allegro_flare/placement2d.h>
 #include <fullscore/models/measure_grid.h>
+#include <fullscore/models/playback_control.h>
 #include <fullscore/services/music_engraver.h>
-#include <fullscore/playback_control.h>
